requirements.txt
aiofiles==24.1.0
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
filelock==3.16.1
crawl4ai==0.4.247
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
logfire==3.1.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
postgrest==0.19.1
protobuf==5.29.3
psutil==6.1.1
pycparser==2.22
pydantic==2.10.5
pydantic-ai==0.0.18
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.1.0
requests==2.32.3
rich==13.9.4
rsa==4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
storage3==0.11.0
streamlit==1.41.1
supabase==2.11.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
websockets==13.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0